Выделил методы создания экземпляров в сервис провайдер
